+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Added
   * Added example and tests for `terra-worklist-data-grid` to select rows via mouse and keyboard.
+  * Added test to `terra-worklist-data-grid` for sticky column header feature.
 
 * Changed
   * Updated focusable cell example for `terra-worklist-data-grid` to allow text area to be editable.